confocal: move shared methods/properties to base classes
