+++ lumicks/pylake/kymo.py
@@ -2,13 +2,12 @@ import json
 import numpy as np
 import warnings
 
-from .confocal import BaseScan, ConfocalImage, axis_label
-from .detail.image import reconstruct_image_sum, reconstruct_image, save_tiff, ImageMetadata, line_timestamps_image, \
-    seek_timestamp_next_line
+from .detail.confocal import ConfocalImage
+from .detail.image import reconstruct_image_sum, reconstruct_image, line_timestamps_image, seek_timestamp_next_line
 from .detail.timeindex import to_timestamp
 
 
-class Kymo(BaseScan, ConfocalImage):
+class Kymo(ConfocalImage):
     """A Kymograph exported from Bluelake
 
     Parameters
@@ -60,29 +59,13 @@ class Kymo(BaseScan, ConfocalImage):
 
         return Kymo(self.name, self.file, start, stop, self.json)
 
-    def _get_photon_count(self, name):
-        photon_count = getattr(self.file, f"{name}_photon_count".lower())[self.start:self.stop]
-
-        if self._has_incorrect_start(photon_count._src.start, photon_count._src.dt):
-            photon_count = getattr(self.file, f"{name}_photon_count".lower())[self.start:self.stop]
-
-        return photon_count
-
-    def _ordered_axes(self):
-        """Returns axis indices in spatial order"""
-        return sorted(self.json["scan volume"]["scan axes"], key=lambda x: x["axis"])
-
-    @property
-    def has_fluorescence(self) -> bool:
-        return self.json["fluorescence"]
-
-    @property
-    def has_force(self) -> bool:
-        return self.json["force"]
-
-    @property
-    def infowave(self):
-        return self.file["Info wave"]["Info wave"][self.start:self.stop]
+    def _fix_incorrect_start(self):#, timeline_start, timeline_dt):
+        """ Resolve error when confocal scan starts before the timeline information.
+            For kymographs this is recoverable by omitting the first line. """
+        self.start = seek_timestamp_next_line(self.infowave[self.start:])
+        self._cache = {}
+        warnings.warn("Start of the kymograph was truncated. Omitting the truncated first line.",
+                        RuntimeWarning)
 
     def _image(self, color):
         if color not in self._cache:
@@ -90,35 +73,9 @@ class Kymo(BaseScan, ConfocalImage):
             self._cache[color] = reconstruct_image_sum(photon_counts, self.infowave.data, self.pixels_per_line).T
         return self._cache[color]
 
-    def _has_incorrect_start(self, timeline_start, timeline_dt):
-        """Checks whether the scan or kymograph starts before the timeline information. If this is the case, it will
-        lead to an incorrect reconstruction. For kymographs this is recoverable by omitting the first line. For scans
-        it is currently unrecoverable."""
-
-        # Workaround for a bug in the STED delay mechanism which could result in scan start times ending up within
-        # the sample time.
-        if timeline_start - timeline_dt < self.start < timeline_start:
-            self.start = timeline_start
-
-        if timeline_start > self.start:
-            if type(self) == Kymo:
-                self.start = seek_timestamp_next_line(self.infowave[self.start:])
-                self._cache = {}
-                warnings.warn("Start of the kymograph was truncated. Omitting the truncated first line.",
-                              RuntimeWarning)
-                return True
-            else:
-                raise RuntimeError("Start of the scan was truncated. Reconstruction cannot proceed. Did you export the "
-                                   "entire scan time in Bluelake?")
-
     def _timestamps(self, sample_timestamps):
         return reconstruct_image(sample_timestamps, self.infowave.data,
                                  self.pixels_per_line, reduce=np.mean).T
-
-    @property
-    def pixelsize_um(self):
-        """Returns a `List` of axes dimensions in um. The length of the list corresponds to the number of scan axes."""
-        return [axes["pixel size (nm)"] / 1000 for axes in self._ordered_axes()]
 
     @property
     def line_time_seconds(self):
@@ -127,37 +84,6 @@ class Kymo(BaseScan, ConfocalImage):
             return (self.timestamps[0, 1] - self.timestamps[0, 0]) / 1e9
         else:
             raise RuntimeError("Line time is not defined for kymograph with only a single line")
-
-    @property
-    def red_image(self):
-        return self._image("red")
-
-    @property
-    def green_image(self):
-        return self._image("green")
-
-    @property
-    def blue_image(self):
-        return self._image("blue")
-
-    @property
-    def rgb_image(self):
-        color_channels = [getattr(self, f"{color}_image").T for color in ("red", "green", "blue")]
-        return np.stack(color_channels).T
-
-    @property
-    def timestamps(self) -> np.ndarray:
-        """Timestamps for image pixels, not for samples
-
-        The returned array has the same shape as the `*_image` arrays.
-        """
-        # Uses the timestamps from the first non-zero-sized photon channel
-        photon_counts = self.red_photon_count, self.green_photon_count, self.blue_photon_count
-        for photon_count in photon_counts:
-            if len(photon_count) == 0:
-                continue
-            return self._timestamps(photon_count.timestamps)
-        raise RuntimeError("Can't get pixel timestamps if there are no pixels")
 
     def _plot(self, image, **kwargs):
         import matplotlib.pyplot as plt
@@ -178,60 +104,6 @@ class Kymo(BaseScan, ConfocalImage):
         plt.xlabel("time (s)")
         plt.ylabel(r"position ($\mu$m)")
         plt.title(self.name)
-
-    def _plot_color(self, color, **kwargs):
-        from matplotlib.colors import LinearSegmentedColormap
-
-        linear_colormaps = {
-            "red": LinearSegmentedColormap.from_list("red", colors=[(0, 0, 0), (1, 0, 0)]),
-            "green": LinearSegmentedColormap.from_list("green", colors=[(0, 0, 0), (0, 1, 0)]),
-            "blue": LinearSegmentedColormap.from_list("blue", colors=[(0, 0, 0), (0, 0, 1)]),
-        }
-
-        image = getattr(self, f"{color}_image")
-        return self._plot(image, **{"cmap": linear_colormaps[color], **kwargs})
-
-    def plot_red(self, **kwargs):
-        """Plot an image of the red photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to :func:`matplotlib.pyplot.imshow`.
-        """
-        return self._plot_color("red", **kwargs)
-
-    def plot_green(self, **kwargs):
-        """Plot an image of the green photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to :func:`matplotlib.pyplot.imshow`.
-        """
-        return self._plot_color("green", **kwargs)
-
-    def plot_blue(self, **kwargs):
-        """Plot an image of the blue photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to :func:`matplotlib.pyplot.imshow`.
-        """
-        return self._plot_color("blue", **kwargs)
-
-    def plot_rgb(self, **kwargs):
-        """Plot a full rbg kymograph image
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to :func:`matplotlib.pyplot.imshow`.
-        """
-        image = self.rgb_image
-        image = image / np.max(image)
-        return self._plot(image, **kwargs)
 
     def _downsample_channel(self, n, xy, reduce=np.mean):
         force = self.file._get_force(n, xy)
@@ -276,32 +148,9 @@ class Kymo(BaseScan, ConfocalImage):
         force = self._downsample_channel(force_channel[-2], force_channel[-1], reduce=reduce)
         force.plot(**kwargs)
         ax2.set_xlim(ax1.get_xlim())
-
-        m = np.vstack((self.timestamps[0,:3], self.timestamps[-1,:3])).mean(0)
-        print(m.astype(np.int64))
-        print(self.timestamps[0,:3].astype(np.int64))
-        print(force.timestamps[:3])
         
         set_aspect_ratio(ax1, aspect_ratio)
         set_aspect_ratio(ax2, aspect_ratio)
-
-    def save_tiff(self, filename, dtype=np.float32, clip=False):
-        """Save the RGB photon counts to a TIFF image
-
-        Parameters
-        ----------
-        filename : str
-            The name of the TIFF file where the image will be saved.
-        dtype : np.dtype
-            The data type of a single color channel in the resulting image.
-        clip : bool
-            If enabled, the photon count data will be clipped to fit into the desired `dtype`.
-            This option is disabled by default: an error will be raise if the data does not fit.
-        """
-        if self.rgb_image.size > 0:
-            save_tiff(self.rgb_image, filename, dtype, clip, ImageMetadata.from_dataset(self.json))
-        else:
-            raise RuntimeError("Can't export TIFF if there are no pixels")
 
 
 class EmptyKymo(Kymo):
