+++ lumicks/pylake/point_scan.py
@@ -1,6 +1,4 @@
-import json
-
-from .confocal import BaseScan
+from .detail.confocal import BaseScan
 
 
 class PointScan(BaseScan):
@@ -20,17 +18,6 @@ class PointScan(BaseScan):
         Dictionary containing scan-specific metadata.
     """
 
-    def _get_photon_count(self, name):
-        return getattr(self.file, f"{name}_photon_count".lower())[self.start:self.stop]
-
-    @property
-    def has_fluorescence(self) -> bool:
-        return self.json["fluorescence"]
-
-    @property
-    def has_force(self) -> bool:
-        return self.json["force"]
-
     def _plot_color(self, color, **kwargs):
         import matplotlib.pyplot as plt
 
@@ -40,36 +27,6 @@ class PointScan(BaseScan):
         plt.xlabel("time (s)")
         plt.ylabel(r"photon count")
         plt.title(self.name)
-
-    def plot_red(self, **kwargs):
-        """Plot the red photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to `~matplotlib.pyplot.plot`.
-        """
-        self._plot_color("red", **kwargs)
-
-    def plot_green(self, **kwargs):
-        """Plot the red photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to `~matplotlib.pyplot.plot`.
-        """
-        self._plot_color("green", **kwargs)
-
-    def plot_blue(self, **kwargs):
-        """Plot the red photon channel
-
-        Parameters
-        ----------
-        **kwargs
-            Forwarded to `~matplotlib.pyplot.plot`.
-        """
-        self._plot_color("blue", **kwargs)
 
     def plot_rgb(self, **kwargs):
         """Plot all color channels
